Fix metadata encoding of too-big notification messages
If a message body is too large, we omit it from the notification, but we
were not encoding that as json and so the metadata would end up as a
(bt-encoded) dict within the notification rather than a bt-encoded
string containing the json.

This fixes it to use the same json encoding as we use when sending a
message that doesn't exceed the limit.