Android: use high priority for notification-inducing message types
Android normal notifications can induce a long delay before the device
gets woken up, and so we really want to use high notifications when we
can.

This gets complicated, though, because Google has some required
threshold of high priority notifications to actual device notifications,
but we cannot know in advance whether a message might induce a
notification (i.e. new DM) or not (typing indicator, or message in a
conversation with notifications disabled).

And so, for now, we disable high priority for messages that we know
cannot illicit a notification, and try high priority for those that
might, and just hope Android's limits aren't too severe.